Mitigate Collections spec race condition
This is a temporary fix until the state management of the Manage
Collections feature can be reworked.

It appears that clicking 'New collection' and filling in values
before fetching the main list of collections (for the index view)
can lead to conflicts where the state for the newly created collection
gets wiped out.

Example of failure in the video recording for this CI run:
https://concourse.devboclips.net/teams/main/pipelines/boclips/jobs/end-to-end-tests/builds/978